Refacto Doctrine Query Source Iterator
