+++ src/Source/DoctrineODMQuerySourceIterator.php
@@ -14,12 +14,9 @@ declare(strict_types=1);
 namespace Sonata\Exporter\Source;
 
 use Doctrine\ODM\MongoDB\Query\Query;
-use Doctrine\ORM\Internal\Hydration\IterableResult;
 use Sonata\Exporter\Exception\InvalidMethodCallException;
-use Symfony\Component\PropertyAccess\PropertyAccess;
-use Symfony\Component\PropertyAccess\PropertyPath;
 
-final class DoctrineODMQuerySourceIterator implements SourceIteratorInterface
+final class DoctrineODMQuerySourceIterator extends AbstractPropertySourceIterator implements SourceIteratorInterface
 {
     /**
      * @var Query
@@ -27,74 +24,20 @@ final class DoctrineODMQuerySourceIterator implements SourceIteratorInterface
     private $query;
 
     /**
-     * @var IterableResult
-     */
-    private $iterator;
-
-    /**
-     * @var array
-     */
-    private $propertyPaths = [];
-
-    /**
-     * @var PropertyAccess
-     */
-    private $propertyAccessor;
-
-    /**
-     * @var string default DateTime format
-     */
-    private $dateTimeFormat;
-
-    /**
-     * @param Query $query  The Doctrine Query
-     * @param array $fields Fields to export
+     * @param array<string> $fields Fields to export
      */
     public function __construct(Query $query, array $fields, string $dateTimeFormat = 'r')
     {
         $this->query = clone $query;
 
-        $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
-
-        foreach ($fields as $name => $field) {
-            if (\is_string($name) && \is_string($field)) {
-                $this->propertyPaths[$name] = new PropertyPath($field);
-            } else {
-                $this->propertyPaths[$field] = new PropertyPath($field);
-            }
-        }
-
-        $this->dateTimeFormat = $dateTimeFormat;
+        parent::__construct($fields, $dateTimeFormat);
     }
 
     public function current()
     {
         $current = $this->iterator->current();
 
-        $data = [];
-
-        foreach ($this->propertyPaths as $name => $propertyPath) {
-            $data[$name] = $this->getValue($this->propertyAccessor->getValue($current, $propertyPath));
-        }
-
-        $this->query->getDocumentManager()->getUnitOfWork()->detach($current);
-
-        return $data;
-    }
-
-    public function next(): void
-    {
-        $this->iterator->next();
-    }
-
-    public function key()
-    {
-        return $this->iterator->key();
-    }
-
-    public function valid(): bool
-    {
-        return $this->iterator->valid();
+        return $this->getCurrentData($current[0]);
     }
 
     public function rewind(): void
@@ -105,31 +48,5 @@ final class DoctrineODMQuerySourceIterator implements SourceIteratorInterface
 
         $this->iterator = $this->query->iterate();
         $this->iterator->rewind();
-    }
-
-    public function setDateTimeFormat(string $dateTimeFormat): void
-    {
-        $this->dateTimeFormat = $dateTimeFormat;
-    }
-
-    public function getDateTimeFormat(): string
-    {
-        return $this->dateTimeFormat;
-    }
-
-    /**
-     * @return mixed
-     */
-    private function getValue($value)
-    {
-        if (\is_array($value) || $value instanceof \Traversable) {
-            $value = null;
-        } elseif ($value instanceof \DateTimeInterface) {
-            $value = $value->format($this->dateTimeFormat);
-        } elseif (\is_object($value)) {
-            $value = (string) $value;
-        }
-
-        return $value;
     }
 }
